[LIB] Add ability to configure request / response options
Both `HttpRequest` and `HttpResponse` now have a
`.with_options(Options)` method that can be used to change or override
the request / response behaviour on an individual basis.

Now, for the time being the only available option is `json_compare_depth`
which allows one to specify the maximum depth for recursive JSON body
comparisions (defaulting to `4096`).